Zombies and Papers added to the game
Made a change in which after 8 seconds of winning the game, the player will be send to the MenuScene
