+++ TheLastHope/Assets/#NVJOB Simple Water Shaders/Examples Water/Materials/Water Specular.mat
@@ -143,8 +143,8 @@ Material:
     m_Floats:
     - _Albedo2Flow: 0.3
     - _Albedo2Tiling: 2.3
-    - _AlbedoContrast: 1.35
-    - _AlbedoIntensity: 1.55
+    - _AlbedoContrast: 3
+    - _AlbedoIntensity: 1.24
     - _Brightness: 1
     - _BumpScale: 1
     - _Contrast: 1
