+++ TheLastHope/Assets/Nokobot/Modern Guns - Handgun/URP/ModernGuns_Handgun_v1.2_URP.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 21037762408ec1942ae9b93b6b1413e7
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
