+++ TheLastHope/Assets/The last hope/Prefabs/Personajes/Zombie.prefab
@@ -523,7 +523,7 @@ NavMeshAgent:
   m_Enabled: 1
   m_AgentTypeID: 0
   m_Radius: 1
-  m_Speed: 0.75
+  m_Speed: 1.2
   m_Acceleration: 2
   avoidancePriority: 50
   m_AngularSpeed: 120
