--- TheLastHope/Assets/The last hope/Prefabs/Personajes/Zombie1.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 86dbeaa434af15a4eb50f2ca7a4a270b
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
