+++ TheLastHope/Assets/The last hope/Scenes/GameScene/NavMesh-NavMesh 1.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 49a59a0b747b9da40a3623113e51533c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
