+++ TheLastHope/Assets/The last hope/Scripts/HealthBarController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d643422b74cf4c84a87d888d0a9c327c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
