+++ TheLastHope/Assets/The last hope/Scripts/gameOver.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5810bb2cef207604b9c009b00b8e722a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
