+++ TheLastHope/ProjectSettings/TagManager.asset
@@ -20,7 +20,7 @@ TagManager:
   - WhatIsLadder
   - WhatIsPlayer
   - WhatIsEnemy
-  - 
+  - HotelFloor
   - 
   - 
   -
